Extract "dependency-version-checker" file
